requirements.txt
numpy>=1.22.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pandas>=1.3.0
pyyaml>=6.0
ultralytics>=8.0.0
torch>=1.13.0
scikit-learn>=1.0.0
albumentations>=1.0.0
folium>=0.12.0
requests>=2.25.0
python-dotenv>=0.19.0 